Unit-tested implementaton of dynamic integers, indev for said references in a Message routine, refactor of the Routine hierarchy for the API, and added a reload/register hook used by implementing the ModDamagePlugin interface. Merged some routines with similar functions.

+++ src/com/KoryuObihiro/bukkit/ModDamage/Backend/IntegerMatching/EntityMatch.java
@@ -61,5 +61,5 @@ public class EntityMatch extends IntegerMatch
 	}
 	
 	@Override
-	public long getValue(TargetEventInfo eventInfo){ return propertyMatch.getProperty(eventInfo, reference);}
+	public int getValue(TargetEventInfo eventInfo){ return propertyMatch.getProperty(eventInfo, reference);}
 }
